requirements.txt
crewai[tools]
streamlit
openai
python-dotenv
requests
beautifulsoup4
langchain
langchain-openai
pydantic